[Edit] Photo 패키지 종속 분리, HairPhoto, ReviewPhoto 분리
