+++ src/main/java/com/inq/wishhair/wesharewishhair/domain/photo/HairPhoto.java
@@ -1,0 +1,26 @@
+package com.inq.wishhair.wesharewishhair.domain.photo;
+
+import com.inq.wishhair.wesharewishhair.domain.hairstyle.HairStyle;
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class HairPhoto {
+
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "hair_style_id")
+    private HairStyle hairStyle;
+
+    @Column(nullable = false, updatable = false)
+    private String originalFilename;
+
+    @Column(nullable = false, updatable = false)
+    private String storeFilename;
+}
