+++ src/main/java/com/inq/wishhair/wesharewishhair/domain/photo/ReviewPhoto.java
@@ -1,6 +1,6 @@
-package com.inq.wishhair.wesharewishhair.domain.hairstyle.photo;
+package com.inq.wishhair.wesharewishhair.domain.photo;
 
-import com.inq.wishhair.wesharewishhair.domain.hairstyle.HairStyle;
+import com.inq.wishhair.wesharewishhair.domain.review.Review;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -9,14 +9,15 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Photo {
+public class ReviewPhoto {
 
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "hair_style_id")
-    private HairStyle hairStyle;
+    @JoinColumn(name = "review_id")
+    private Review review;
 
     @Column(nullable = false, updatable = false)
     private String originalFilename;
@@ -24,3 +25,4 @@ public class Photo {
     @Column(nullable = false, updatable = false)
     private String storeFilename;
 }
+
